rename #157: DTO 클래스 패키지 구조 변경
