+++ src/main/java/fastcampus/team7/Livable_officener/dto/signup/PhoneAuthDTO.java
@@ -1,4 +1,4 @@
-package fastcampus.team7.Livable_officener.dto;
+package fastcampus.team7.Livable_officener.dto.signup;
 
 import lombok.*;
